refactor(pool-indexer api): split ticks/pools by endpoint, wrap Networks, add HistogramVecExt
