+++ crates/e2e/tests/e2e/pool_indexer.rs
@@ -24,6 +24,7 @@ use {
         MetricsConfig,
         NetworkConfig,
         NetworkName,
+        Networks,
     },
     serde::Deserialize,
     sqlx::PgPool,
@@ -194,7 +195,7 @@ async fn spawn_pool_indexer(factory: Address, metrics_port: u16) -> tokio::task:
             url: LOCAL_DB_URL.parse().unwrap(),
             max_connections: NonZeroU32::new(5).unwrap(),
         },
-        networks: vec![NetworkConfig {
+        networks: Networks::try_new(vec![NetworkConfig {
             name: NetworkName::new("mainnet"),
             chain_id: 1,
             rpc_url: "http://127.0.0.1:8545".parse().unwrap(),
@@ -206,7 +207,8 @@ async fn spawn_pool_indexer(factory: Address, metrics_port: u16) -> tokio::task:
             seed_block: None,
             fetch_concurrency: 8,
             prefetch_concurrency: 50,
-        }],
+        }])
+        .unwrap(),
         api: ApiConfig {
             bind_address: SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, POOL_INDEXER_PORT)),
         },
