+++ crates/pool-indexer/src/config.rs
@@ -1,7 +1,7 @@
 use {
     alloy_primitives::Address,
     anyhow::{Context, Result},
-    serde::Deserialize,
+    serde::{Deserialize, Deserializer},
     std::{
         collections::HashSet,
         fmt,
@@ -186,7 +186,7 @@ impl Default for MetricsConfig {
 pub struct Configuration {
     pub database: DatabaseConfig,
     #[serde(rename = "network")]
-    pub networks: Vec<NetworkConfig>,
+    pub networks: Networks,
     #[serde(default)]
     pub api: ApiConfig,
     #[serde(default)]
@@ -197,38 +197,76 @@ impl Configuration {
     pub fn from_path(path: &Path) -> Result<Self> {
         let content = std::fs::read_to_string(path)
             .with_context(|| format!("reading config file {}", path.display()))?;
-        let config: Self = toml::from_str(&content).context("parsing config file")?;
-        config.validate_networks()?;
-        Ok(config)
+        toml::from_str(&content).context("parsing config file")
+    }
+}
+
+/// Validated list of [`NetworkConfig`]. Construction enforces cross-network
+/// uniqueness (names and chain_ids) and the exactly-one-factory invariant —
+/// so the rest of the codebase can iterate without re-checking.
+#[derive(Debug)]
+pub struct Networks(Vec<NetworkConfig>);
+
+impl Networks {
+    pub fn try_new(networks: Vec<NetworkConfig>) -> Result<Self> {
+        validate_networks(&networks)?;
+        Ok(Self(networks))
     }
 
-    /// Cross-network sanity checks that don't fit serde's per-field validation:
-    /// uniqueness of names / chain IDs and the exactly-one-factory invariant.
-    fn validate_networks(&self) -> Result<()> {
-        anyhow::ensure!(
-            !self.networks.is_empty(),
-            "at least one [[network]] must be configured",
-        );
-        let mut names = HashSet::new();
-        let mut chain_ids = HashSet::new();
-        for n in &self.networks {
-            anyhow::ensure!(
-                names.insert(n.name.as_str()),
-                "duplicate network name: {}",
-                n.name,
-            );
-            anyhow::ensure!(
-                chain_ids.insert(n.chain_id),
-                "duplicate chain_id: {}",
-                n.chain_id,
-            );
-            anyhow::ensure!(
-                n.factories.len() == 1,
-                "network {}: exactly one factory per network is supported in this release, got {}",
-                n.name,
-                n.factories.len(),
-            );
-        }
-        Ok(())
+    pub fn as_slice(&self) -> &[NetworkConfig] {
+        &self.0
     }
+}
+
+impl IntoIterator for Networks {
+    type IntoIter = std::vec::IntoIter<NetworkConfig>;
+    type Item = NetworkConfig;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.into_iter()
+    }
+}
+
+impl<'a> IntoIterator for &'a Networks {
+    type IntoIter = std::slice::Iter<'a, NetworkConfig>;
+    type Item = &'a NetworkConfig;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.iter()
+    }
+}
+
+impl<'de> Deserialize<'de> for Networks {
+    fn deserialize<D: Deserializer<'de>>(de: D) -> Result<Self, D::Error> {
+        let networks = Vec::<NetworkConfig>::deserialize(de)?;
+        Self::try_new(networks).map_err(serde::de::Error::custom)
+    }
+}
+
+fn validate_networks(networks: &[NetworkConfig]) -> Result<()> {
+    anyhow::ensure!(
+        !networks.is_empty(),
+        "at least one [[network]] must be configured",
+    );
+    let mut names = HashSet::new();
+    let mut chain_ids = HashSet::new();
+    for n in networks {
+        anyhow::ensure!(
+            names.insert(n.name.as_str()),
+            "duplicate network name: {}",
+            n.name,
+        );
+        anyhow::ensure!(
+            chain_ids.insert(n.chain_id),
+            "duplicate chain_id: {}",
+            n.chain_id,
+        );
+        anyhow::ensure!(
+            n.factories.len() == 1,
+            "network {}: exactly one factory per network is supported in this release, got {}",
+            n.name,
+            n.factories.len(),
+        );
+    }
+    Ok(())
 }
